get_phone_numbers_from_krr/requirements.txt
--extra-index-url https://europe-west3-python.pkg.dev/innbyggerkontakt-dev/ik-python-repo/simple
--extra-index-url https://europe-west3-python.pkg.dev/innbyggerkontakt/ik-python-repo/simple

cachecontrol==0.12.11 ; python_version >= "3.10" and python_version < "4.0"
cachetools==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
certifi==2022.9.24 ; python_version >= "3.10" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.10" and python_version < "4.0"
charset-normalizer==2.1.1 ; python_version >= "3.10" and python_version < "4"
click==8.1.3 ; python_version >= "3.10" and python_version < "4"
cloud-event-model==0.1.4 ; python_version >= "3.10" and python_version < "4.0"
cloudevents==1.6.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.5 ; python_version >= "3.10" and python_version < "4" and platform_system == "Windows"
cryptography==36.0.1 ; python_version >= "3.10" and python_version < "4.0"
deprecated==1.2.13 ; python_version >= "3.10" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.10" and python_version < "4.0"
dnspython==2.2.1 ; python_version >= "3.10" and python_version < "4.0"
email-validator==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
firebase-admin==5.2.0 ; python_version >= "3.10" and python_version < "4.0"
flask==2.2.2 ; python_version >= "3.10" and python_version < "4"
freg-models==0.1.1 ; python_version >= "3.10" and python_version < "4.0"
functions-framework==3.2.0 ; python_version >= "3.10" and python_version < "4"
google-api-core==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
google-api-core[grpc]==2.10.1 ; python_version >= "3.10" and python_version < "4.0"
google-api-python-client==2.64.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth==2.12.0 ; python_version >= "3.10" and python_version < "4.0"
google-auth-httplib2==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-firestore==2.7.1 ; python_version >= "3.10" and python_version < "4.0" and platform_python_implementation != "PyPy"
google-cloud-pubsub==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-secret-manager==2.9.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-storage==1.43.0 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-tasks==2.7.1 ; python_version >= "3.10" and python_version < "4.0"
google-cloud-trace==1.7.2 ; python_version >= "3.10" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.10" and python_version < "4.0"
google-events==0.2.0 ; python_version >= "3.10" and python_version < "4.0"
google-resumable-media==2.4.0 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos==1.56.4 ; python_version >= "3.10" and python_version < "4.0"
googleapis-common-protos[grpc]==1.56.4 ; python_version >= "3.10" and python_version < "4.0"
grpc-google-iam-v1==0.12.4 ; python_version >= "3.10" and python_version < "4.0"
grpcio==1.49.1 ; python_version >= "3.10" and python_version < "4.0"
grpcio-status==1.49.1 ; python_version >= "3.10" and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.10" and python_version < "4" and platform_system != "Windows"
httplib2==0.20.4 ; python_version >= "3.10" and python_version < "4.0"
idna==3.4 ; python_version >= "3.10" and python_version < "4"
innbyggerkontakt==1.0.2 ; python_version >= "3.10" and python_version < "4.0"
itsdangerous==2.1.2 ; python_version >= "3.10" and python_version < "4"
jinja2==3.1.2 ; python_version >= "3.10" and python_version < "4"
krr-models==0.1.0 ; python_version >= "3.10" and python_version < "4.0"
libcst==0.4.7 ; python_version >= "3.10" and python_version < "4.0"
markupsafe==2.1.1 ; python_version >= "3.10" and python_version < "4"
msgpack==1.0.4 ; python_version >= "3.10" and python_version < "4.0"
mypy-extensions==0.4.3 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-api==1.13.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-exporter-gcp-trace==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-flask==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-requests==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-instrumentation-wsgi==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-propagator-gcp==1.3.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-sdk==1.13.0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-semantic-conventions==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
opentelemetry-util-http==0.34b0 ; python_version >= "3.10" and python_version < "4.0"
otel-setup==0.2.3 ; python_version >= "3.10" and python_version < "4.0"
packaging==21.3 ; python_version >= "3.10" and python_version < "4.0"
phone-validator-models==0.1.2 ; python_version >= "3.10" and python_version < "4.0"
proto-plus==1.22.1 ; python_version >= "3.10" and python_version < "4.0"
protobuf==4.21.7 ; python_version >= "3.10" and python_version < "4.0"
ps-message-krr-message==0.1.1 ; python_version >= "3.10" and python_version < "4.0"
ps-message-sms-event==0.1.1 ; python_version >= "3.10" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.10" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.10" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.10" and python_version < "4.0"
pydantic==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pydantic[email]==1.10.2 ; python_version >= "3.10" and python_version < "4.0"
pyjwt==2.3.0 ; python_version >= "3.10" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.10" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.10" and python_version < "4.0"
pytz==2022.1 ; python_version >= "3.10" and python_version < "4.0"
pyyaml==6.0 ; python_version >= "3.10" and python_version < "4.0"
requests==2.28.1 ; python_version >= "3.10" and python_version < "4"
rsa==4.9 ; python_version >= "3.10" and python_version < "4"
setuptools==65.4.1 ; python_version >= "3.10" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.10" and python_version < "4.0"
typing-extensions==4.3.0 ; python_version >= "3.10" and python_version < "4.0"
typing-inspect==0.8.0 ; python_version >= "3.10" and python_version < "4.0"
uritemplate==4.1.1 ; python_version >= "3.10" and python_version < "4.0"
urllib3==1.26.12 ; python_version >= "3.10" and python_version < "4"
watchdog==2.1.9 ; python_version >= "3.10" and python_version < "4"
werkzeug==2.2.2 ; python_version >= "3.10" and python_version < "4"
wrapt==1.14.1 ; python_version >= "3.10" and python_version < "4.0"
